refactor(macros): split DIContainerValidator helpers into dedicated files (N-1)
Extracts ~490 lines of diagnostic, fix-it, and type-check helpers from
`DIContainerValidator.swift` (904 → 418) without touching the
validator's core `validate(model:context:)` loop. Macro snapshots and
diagnostic-code tests stay byte-identical; all 349 tests pass.

New files:

- `DIProvideValidationDiagnostics.swift` — owns every `SimpleDiagnostic`
  / `SimpleNote` / `FixIt` constructor the validator reaches for:
  `makeUnresolvedFactoryParameterDiagnostic`,
  `makeUnresolvedWithDependencyDiagnostic`,
  `makeUnavailableDependencyDiagnostic`, `makeConcreteOptInDiagnostic`,
  plus the `DirectProviderEagerCallSite` detection used by the Phase L
  Provider eager-call check, `matchingDependencyCandidates` suggestions,
  and the generic rename/replace/concrete fix-it builders. Keeping the
  diagnostic message / fix-it logic together lets Phase N-3 (child-
  overrides-missing) and N-4 (Lazy/Provider alias warning) extend this
  file without touching the validator itself.
- `DIContainerValidatorTypeChecks.swift` — type-shape primitives used
  by the validator: `requiresConcreteOptIn`, `normalizedConcreteCheckType`,
  `isExistentialIdentifier`, `isAsyncClosureExpression`. These are
  purely syntactic helpers with no diagnostic state and belong together.

Helpers are promoted from `private` to `internal` where the validator
(still in its original file) needs to call them; no `public` API change.
No behaviour changes.
