+++ Sources/InnoDIMacros/DIContainerValidatorTypeChecks.swift
@@ -1,0 +1,74 @@
+//
+//  DIContainerValidatorTypeChecks.swift
+//  InnoDIMacros
+//
+//  Phase N-1 — extracted from `DIContainerValidator.swift`.
+//
+//  Type-level helpers the validator leans on when deciding whether a
+//  `@Provide`d member needs `concrete: true`, when normalizing optional /
+//  tuple / attributed wrappers, and when detecting async closures. Pulling
+//  these out keeps `DIContainerValidator.swift` focused on the per-member
+//  validation loop itself.
+//
+
+import SwiftSyntax
+
+internal func requiresConcreteOptIn(type: TypeSyntax) -> Bool {
+    let normalized = normalizedConcreteCheckType(type)
+
+    if normalized.is(SomeOrAnyTypeSyntax.self) || normalized.is(CompositionTypeSyntax.self) {
+        return false
+    }
+
+    if let identifier = normalized.as(IdentifierTypeSyntax.self) {
+        return !isExistentialIdentifier(identifier.name.text)
+    }
+
+    if let member = normalized.as(MemberTypeSyntax.self) {
+        return !isExistentialIdentifier(member.name.text)
+    }
+
+    return true
+}
+
+private func normalizedConcreteCheckType(_ type: TypeSyntax) -> TypeSyntax {
+    if let optional = type.as(OptionalTypeSyntax.self) {
+        return normalizedConcreteCheckType(optional.wrappedType)
+    }
+
+    if let implicitlyUnwrapped = type.as(ImplicitlyUnwrappedOptionalTypeSyntax.self) {
+        return normalizedConcreteCheckType(implicitlyUnwrapped.wrappedType)
+    }
+
+    if let attributed = type.as(AttributedTypeSyntax.self) {
+        return normalizedConcreteCheckType(attributed.baseType)
+    }
+
+    if let tuple = type.as(TupleTypeSyntax.self),
+       tuple.elements.count == 1,
+       let first = tuple.elements.first,
+       first.firstName == nil,
+       first.secondName == nil {
+        return normalizedConcreteCheckType(first.type)
+    }
+
+    if let identifier = type.as(IdentifierTypeSyntax.self),
+       identifier.name.text == "Optional",
+       let wrapped = identifier.genericArgumentClause?.arguments.first?.argument,
+       let wrappedType = wrapped.as(TypeSyntax.self) {
+        return normalizedConcreteCheckType(wrappedType)
+    }
+
+    return type
+}
+
+private func isExistentialIdentifier(_ name: String) -> Bool {
+    name == "Any" || name == "AnyObject"
+}
+
+internal func isAsyncClosureExpression(_ expr: ExprSyntax) -> Bool {
+    guard let closure = expr.as(ClosureExprSyntax.self) else {
+        return false
+    }
+    return closure.signature?.effectSpecifiers?.asyncSpecifier != nil
+}
